audio/whisper/requirements.txt
accelerate==0.24.0
jiwer==3.0.3
pyctcdecode==0.5.0
torcheval==0.0.6
librosa==0.10.1
torchaudio==2.1.0
transformers==4.35.2
datasets==2.15.0
lightning==2.1.0
bnunicodenormalizer==0.1.6
evaluate==0.4.1